spanify: Handle `subspan()` in binary operations
Coerce `subspan()` arguments as necessary when emitting these for binary
operators (e.g., `a + b`).

Bug: 412395846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6634109
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1474235}

+++ tools/clang/spanify/tests/return-stmts-expected.cc
@@ -2,9 +2,11 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // Copyright 2024 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
+#include <cstdint>
 #include <vector>
 
 #include "base/containers/span.h"
+#include "base/numerics/safe_conversions.h"
 
 // Expected rewrite:
 // base::span<int> fct1()
@@ -56,7 +58,9 @@ base::span<int> fct5() {
   // base::span<int> var1 = new int[1024];
   base::span<int> var1 = new int[1024];
   int offset = 1;
-  return var1.subspan(offset);
+  // Expected rewrite:
+  // return var1.subspan(base::checked_cast<size_t>(offset));
+  return var1.subspan(base::checked_cast<size_t>(offset));
 }
 
 // Expected rewrite:
@@ -68,8 +72,10 @@ base::span<char> fct6() {
   base::span<int> var1 = new int[1024];
   int offset = 1;
   // Expected rewrite:
-  // return base::as_writable_byte_span(var1.subspan(offset));
-  return base::as_writable_byte_span(var1.subspan(offset));
+  // return
+  // base::as_writable_byte_span(var1.subspan(base::checked_cast<size_t>(offset)));
+  return base::as_writable_byte_span(
+      var1.subspan(base::checked_cast<size_t>(offset)));
 }
 
 // Function return type not rewritten since not used.
@@ -80,8 +86,8 @@ int* fct7() {
   base::span<int> var1 = new int[1024];
   int offset = 1;
   // Expected rewrite:
-  // return var1.subspan(offset).data();
-  return var1.subspan(offset).data();
+  // return var1.subspan(base::checked_cast<size_t>(offset)).data();
+  return var1.subspan(base::checked_cast<size_t>(offset)).data();
 }
 
 // Function return type not rewritten since not used.
@@ -92,10 +98,13 @@ char* fct8() {
   base::span<int> var1 = new int[1024];
   int offset = 1;
   // Expected rewrite:
-  // return reinterpret_cast<char*>(var1).subspan(offset).data();
+  // return
+  // reinterpret_cast<char*>(var1).subspan(base::checked_cast<size_t>(offset)).data();
   // As-is, this code doesn't compile because we don't yet handle
   // adapting these reinterpret_cast expressions for spans.
-  return reinterpret_cast<char*>(var1).subspan(offset).data();
+  return reinterpret_cast<char*>(var1)
+      .subspan(base::checked_cast<size_t>(offset))
+      .data();
 }
 
 void usage() {
